avd2/requirements.txt
pulumi>=3
pulumi_azure_native.resources
pulumi_azure_native.desktopvirtualization